fixes #596 make sure if and is helper get passed observables, not values
